backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
lightrag-hku
google-genai
numpy
pydantic
